Fixed a Linux warning.
git-svn-id: http://svn.osgeo.org/fdo/trunk@3683 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/GenericRdbms/Src/Gdbi/GdbiQueryResult.cpp
@@ -686,8 +686,8 @@ GdbiColumnInfoType *GdbiQueryResult::FindColumnCache( const wchar_t *colName )
 		FdoStringP  upperName = FdoStringP(colName).Upper();
 		const wchar_t* name = StripTable(upperName);
 
-        size_t index_empty = -1;
-		for ( size_t i = 0; !found && i < mColList->size(); i++ )
+        int     index_empty = -1;
+		for ( int i = 0; !found && i < (int)mColList->size(); i++ )
 		{
 			colInfo = mColList->at(i);
 			found = ( wcscmp( colInfo->name, name ) == 0 );
